test: split experimental config tests
Keep config core tests below the file-size guardrail while preserving
coverage for the Azure DevOps experimental opt-in gate.

Closes #404
